requirements.txt
flask
flask-sqlalchemy
flask-login
flask-socketio
flask-migrate
werkzeug
openpyxl
gunicorn
psycopg2-binary
python-dotenv
eventlet
